Fix math being done in int space for assignment to long
* re #5122

+++ enderio-conduits/src/main/java/crazypants/enderio/powertools/machine/capbank/network/CapBankNetwork.java
@@ -155,7 +155,7 @@ public class CapBankNetwork implements ICapBankNetwork, ServerTickHandler.ITickL
   public void addMember(@Nonnull TileCapBank cap) {
     if (!capBanks.contains(cap)) {
       capBanks.add(cap);
-      long newIO = maxIO + cap.getType().getMaxIO();
+      long newIO = (long) maxIO + cap.getType().getMaxIO();
       if (newIO > IO_CAP) {
         newIO = IO_CAP;
       }
